+++ Pod/Classes/Custom Modules/MPCustomModule.h
@@ -18,10 +18,12 @@
 
 #import <Foundation/Foundation.h>
 
+@class MPCustomModulePreference;
+
 @interface MPCustomModule : NSObject <NSCopying, NSCoding>
 
 @property (nonatomic, strong, readonly, nonnull) NSNumber *customModuleId;
-@property (nonatomic, strong, readonly, nullable) NSArray *preferences;
+@property (nonatomic, strong, readonly, nullable) NSArray<MPCustomModulePreference *> *preferences;
 
 - (nonnull instancetype)initWithDictionary:(nonnull NSDictionary *)customModuleDictionary;
 - (nonnull NSDictionary *)dictionaryRepresentation;