+++ Pod/Classes/Data Model/MPConsumerInfo.h
@@ -43,7 +43,7 @@ extern NSString * _Nonnull const kMPCKExpiration;
 @interface MPConsumerInfo : NSObject <NSCoding>
 
 @property (nonatomic, unsafe_unretained) int64_t consumerInfoId;
-@property (nonatomic, strong, nullable) NSArray *cookies;
+@property (nonatomic, strong, nullable) NSArray<MPCookie *> *cookies;
 @property (nonatomic, strong, nonnull) NSNumber *mpId;
 @property (nonatomic, strong, nullable) NSString *uniqueIdentifier;