+++ Pod/Classes/Data Model/MPSegment.h
@@ -18,6 +18,8 @@
 
 #import "MPDataModelAbstract.h"
 
+@class MPSegmentMembership;
+
 extern NSString *const kMPSegmentListKey;
 
 @interface MPSegment : MPDataModelAbstract <NSCopying>
@@ -26,10 +28,10 @@ extern NSString *const kMPSegmentListKey;
 @property (nonatomic, strong) NSArray *endpointIds;
 @property (nonatomic, strong, readonly) NSDate *expiration;
 @property (nonatomic, strong) NSString *name;
-@property (nonatomic, strong) NSArray *memberships;
+@property (nonatomic, strong) NSArray<MPSegmentMembership *> *memberships;
 @property (nonatomic, unsafe_unretained, readonly) BOOL expired;
 
-- (instancetype)initWithSegmentId:(NSNumber *)segmentId UUID:(NSString *)uuid name:(NSString *)name memberships:(NSArray *)memberships endpointIds:(NSArray *)endpointIds;
+- (instancetype)initWithSegmentId:(NSNumber *)segmentId UUID:(NSString *)uuid name:(NSString *)name memberships:(NSArray<MPSegmentMembership *> *)memberships endpointIds:(NSArray *)endpointIds;
 - (instancetype)initWithDictionary:(NSDictionary *)segmentDictionary;
 
 @end